change123

+++ sensor/hdi_service/sensor_callback_vdi.cpp
@@ -27,7 +27,7 @@ namespace V2_0 {
 namespace {
     constexpr int32_t DATA_LEN = 256;
     constexpr int64_t REPOPRT_TIME = 60000000000;
-    constexpr int32_t INIT_DATA_COUNT = 1;
+    constexpr int64_t INIT_DATA_COUNT = 1;
     static std::unordered_map<int32_t, int64_t> firstTimestampMap_;
     static std::unordered_map<int32_t, int64_t> lastTimestampMap_;
 }
@@ -63,7 +63,7 @@ void SensorCallbackVdi::PrintData(const HdfSensorEvents &event, const std::strin
     std::unique_lock<std::mutex> lock(timestampMapMutex_);
     static std::unordered_map<int32_t, int64_t> sensorDataCountMap;
     auto it = sensorDataCountMap.find(event.sensorId);
-    int32_t dataCount = INIT_DATA_COUNT;
+    int64_t dataCount = INIT_DATA_COUNT;
     if (it == sensorDataCountMap.end()) {
         sensorDataCountMap[event.sensorId] = INIT_DATA_COUNT;
     } else {
